requirements.txt
asgiref
astunparse
certifi
chardet
Django
httplib2
idna
jbxapi
pip
pycryptodomex
pyparsing
PySimpleSOAP
pytz
pyzipper
sandboxapi
setuptools
six
sqlparse
urllib3
xmltodict
reportbug
whitenoise
pycrypto